Add actual run time timers to runners
Note: running using tcc_run doesn't give me a hook point to start the timer
without including the compile time. So not for that one.

Also, this depends on gettimeofday actually having useful accuracy.
Generally, if the accuracy is poor, (eg worse than a few hundred
microseconds per step) the feature to exclude the I/O won't work
because it will appear that the I/O always takes no time.