Time to look at powershell scripting.
Okay, it seems you are "not supposed to used [Console]::Write".
Right, the "correct" thing to use is "Write-Output"
... Um, "Flush-Output" ? "Read-Input" ?

Sigh. It seems that "Write-Output" has nothing to do with character
streams or interaction with the user. It's purely for saving objects to
pass to another code fragment. (Often the 'Convert to text and display'
fragment)

Looks like I need to use "Read-Host" and "Write-Host", but it seems they
also do odd things to the console and break ansicon ... sometimes.

Well, a line-buffered "Write-Host" seems to work for hanoi.b

Now, how to run it? Double click ... nope, Right click & run ... nope,
Right click & edit & run ... NO!! ... Google ... WTF!!!! ... FINE!
Save it as a BAT file and copy&paste a really fucked up bit of polyglot
scripting at the top. NOW it works!

Oh, and one last thing ... appending a ": " to the prompt string, that's
a bug from the sixties!