Updates to DynASM support.
The Makefile now explicitly checks, using the C compiler, for supported
processors. A compiler that understands the GCC option -dM is required
for the detection to work.

Both DynASM and the built in GNU Lightning will only be compiled if
a supported CPU is found.

I addition with the update of the DynASM assembler it's no longer
necessary to have a distinct C file just to put the #include lines
in so the compile itself is a little simpler.