Recovered this hellobox 103 from an old commit.
Looks like the Hello World! was written late in 2014.  But the border
was improved somewhat later.  The combination is very good at detecting
interpreters that don't work exactly as they are expected to.

Note that an interpreter with cells limited to 0..255 will not be able to
run this, however, big number cells and seven bit wrapping cells are fine.