Replacement for open_memstream on "old crap" systems.
It uses mktemp(), but that's ok because it'll never be used anyway.

This is also used as a Python library bug workaround.

The Python library sets a particular value of the XOPEN and POSIX macros.

Leaving aside the facts that it should be touching them at all and
anything that does touch them should only set one of them. It causes
problems because this will be a request to disable later facilities from
the C library.