Adjust the -fwrapv detector to warn with -Wall.
GCC will now warn about the overflowing code with -Wstrict-overflow=1
which is included in -Wall.  Also provide a message in the help to
the effect that we have noted that an old GCC is making very simple
'optimisations' that assume integers don't overflow despite that fact
that they are probably to dumb to break anything.

Note: this detector still does NOT trigger an abort with -ftrapv despite
causing an overflow.