Rename bf2bash as bf2ksh
It seems the generated code is closer to original ksh style that bash
style and works happliy as either if the I/O is okay. So I'm defaulting
to I/O that works with both and only having bash I/O if the bash option
is added.

The bash only I/O is unchanged and uses no external programs.