Add Brainfuck to Whitespace
Also "semicolon" is a Whitespace clone.

Note:

Some WS interpreters think that every cell must be set before use so by
default I add a loop to the start of the program that does that for the
defined tape.

I should probably add the code that automatically zeros the next cell
(if required) at some point as the reference interpreter does a similar
thing where only cells (at or) below the highest written to are allocated.