Workaround for whitespace signed label bug.
Several interpreters treat the "number" for labels as a signed integer
in the same format as "push". This allows that bug by to be worked around
by adding the extra space.

Note: I already add an unneeded tab (one bit) at the start of the "number"
because *a lot* of interpreters treat the label as an unsigned integer
where leading zeros are not significant.