Alter the regexs for recognising Ook commands.
The two Ooks in a command are matched as one unit but must be delimited by
a word boundary.  One exception to this is that recent research implies
that The Librarian may have been saying 'Book' not 'Ook' so this variant
is also accepted.

The "Ook? Ook?" command is the debugging command that this interpreter
does not support.