Okay, checking 30k cells is too slow.
The 18 million steps is too many for some interesting interpreters so
I've moved the check to it's own script adjusting it along the way so
that it checks exactly 30000 cells.

Bitwidth now uses 131 cells (from the check for a positive move of 128)
and runs for 79627 steps on an 8bit interpreter.