Update txtbf, add lookahead processing.
Lookahead processing is expensive so it's been reduced to look only a
short distance ahead and only done (by default) on the results that have
been 'found'. With these changes it can be done by default without too
much expense.

Also some help page updates and a tweak to the 'zoned' processing
for duplicate cells.