Update to interface between FE and BE
I've introduced some common variables that are, optionally, set by
the backend.  Most of the time they are not set and the FE uses it's
defaults. For the simple options, switching optimisations and cell
sizes, the BE only has to redefine the relevent variable with a set
flag. Additional arguments use an optional function.