Disable an optimisation in Tritium.
An implementation of Towers of Hanoi By Jo King has proven that this
optimisation does not in fact preserve the code meaning. Just disable
it for now, try and work out the missed precondition later. Right now
it looks like the problem is that there is a collision between input
and output cell numbers.