Improve movement merge processing.
The aim was to move it into bf2const. But it became very small, small
enough that the support code needed in the BE for an implementation in
bf2const probably wouldn't be any smaller.

So I tried extending it to the full offset processing as in Tritium.
This was added to bf2crun, bf2dump, bf2ruby and bf2py