Supertree rewrite
The old manual Supertree routines try to allocate the entire tape in
a single malloc. This tends not to work very well in part because the
optimiser will cause us to allocate cells that won't actually ever be
used. This routine now only allocates the memory for a cell the first
time that it's modified.